audit: fix all 24 findings of the 2026-07-03 multitenant code audit
Squash of the 19-commit audit-fixes branch. Full report with per-finding
validation evidence: docs/superpowers/diagnostics/2026-07-03-multitenant-
code-audit.md (gitignored). Every fix was validated on the cluster512
farm cells by clean build, the regression suite, and jobs.csv
bit-identity, with exactly two intentional behavior changes (noted
below).

Correctness:
- P0-1/P0-2/P4-11: per-job (rfold/folding) wiring no longer truncates
  in-flight routes. Chunks transiting a rewired device were "refreshed"
  toward their next hop and delivered there as if arrived, silently
  dropping transit latency (130k misdeliveries per 200-job rho0.5 rfold
  cell). Stale refresh now targets the true destination and scoped
  wiring no longer bumps per-device iteration counters, force-frees
  busy unchanged links, or aborts on a busy changed link.
  BEHAVIOR CHANGE (the point of the fix): historical rfold/folding JCT
  tails were optimistic -- measured p90 +0.9%, p99 +5.1%, max +8.6% on
  the validation cell; medians survive. rfold/folding sweep arms need
  re-running.
- P0-3/P1-8: stream ids confined to the collective tag window (windowed
  base + static_assert); previously overflowed int from job_id 108.
- P1-1: job OCS wiring/route overrides torn down at completion via
  ReconfigHook::release (previously grew monotonically forever).
- P1-2: pg-less collectives fail fast in scheduled mode (null deref,
  then a wedge via dims-fallback enrolling non-member NPUs).
- P1-3: issue_dep_free_nodes runs to a fix-point; synchronously
  completing nodes no longer strand a rank until global quiescence.
- P1-4: comm_group.json ranks validated at parse (heap overflow under
  topomatch).
- P1-7: cross-rank comm-group ordinal agreement verified at attach via
  per-CG fingerprints.
- P4 batch: fail-fast on malformed inputs, unknown CLI flags rejected,
  statistics survive real traces, duplicate-NPU placement guard,
  inactive-DataSet fail-fast, deterministic scatter ordering, 0-BW
  link recheck, per-NPU scheduler state reset at tenant detach.
  BEHAVIOR CHANGE: deterministic backfill-reservation tie-break
  (projected_end, ranks, job_id) shifted 7/200 rows on one easy/rfold
  validation cell.

Memory:
- P2 batch: dead write-only static maps deleted (BaseStream
  synchronizer, usage tracking, start_times), chunk-id and node-id maps
  pruned at retire (423k -> 1.3k peak entries), pg-metadata overwrite
  leak, CustomAlgorithm feeder leak.

Performance (validation cells: -24% wall clock):
- P3 batch: event queue linear scan -> time-keyed map, running-jobs
  list for EASY backfill, sched-context built only for rankers that
  want it, node->cg map instead of per-check feeder attribute fetch,
  verbose-logging guards, matrix rows by const&, redundant constructor
  trace parse dropped, resettable dependency resolver (iteration rewind
  instead of re-parse).

Tests: tests/rt_sync_node_strand (P1-3) and tests/rt_ordinal_divergence
(P1-7) added to the regression suite; both synthesize their Chakra
traces from scratch.

Dead code: Scheduler, CSVWriter, UsageTracker, BaseStream stream-sync
statics, Link::get_current_time, unused Chunk::src_id/dest_id, stray
glibc <link.h> include.

Audit corrections recorded in the report: P1-5 refuted (asserts are
LIVE in Release builds -- no NDEBUG), P1-6 downgraded, P4-13 refuted
(suggested fix was a double-free; mixed ownership documented instead).

+++ astra-sim/system/DataSet.cc
@@ -10,7 +10,7 @@ LICENSE file in the root directory of this source tree.
 
 using namespace AstraSim;
 
-int DataSet::id_auto_increment = 0;
+uint64_t DataSet::id_auto_increment = 0;
 
 DataSet::DataSet(int total_streams) {
     this->my_id = id_auto_increment++;
